pyproject.toml
[project]
name = "Nameit"
version = "0.1"
description = "Generate captions for images using GIT from MSFT"
readme = "README.md"
authors = [{name = "Ashvanth.S"}]
license = {text="Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha"
]

dependencies = [
    "transformers",
    "click",
    "torch",
    "pillow",
    "rich",
    "pytest"
]

[project.urls]
Homepage = "https://github.com/ash-01xor/Imgcap"

[project.scripts]
imgcap = "imgcap:cli"

[project.optional-dependencies]
test = ["pytest","pytest-mock"]